% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_census2016.R
\docType{data}
\name{abs_census2016}
\alias{abs_census2016}
\title{Australian Bureau of Statistics Educational Qualifications (2016)}
\format{
A tibble with 64 rows and 4 columns
\describe{
\item{Education Qualification}{Education qualification level based on the Australian Standard Classification of Education (ASCED)}
\item{Territory/State}{Territory or State in Australia}
\item{Count}{Count of persons aged 15 and over}
\item{Percent (\%)}{Percentage of holders with a given education-level as a proportion of total persons in the respective territory/state}
}
}
\source{
Australian Bureau of Statistics (2018) 'Educational Qualifications in Australia', \emph{Census of Population and Housing: Reflecting Australia - Stories from the Census, 2016}, accessed 21 November 2022.

\url{https://www.abs.gov.au/ausstats/abs@.nsf/Lookup/by\%20Subject/2071.0~2016~Main\%20Features~Educational\%20Qualifications\%20Data\%20Summary\%20~65}
}
\usage{
abs_census2016
}
\description{
A subset of data from the 2016 Census of Population and Housing:
Reflecting Australia - Stories from the Census, 2016 - Eudcational
Qualification. Gives data on level of education by territory
of usual residence for persons aged 15 and over.
}
\keyword{datasets}
