% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_partp2017.R
\docType{data}
\name{abs_partp2017}
\alias{abs_partp2017}
\title{Australian Marriage Law Postal Survey Participation Data (2017)}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 31 rows and 17 columns.
}
\source{
Australian Bureau of Statistics (2017) 'Australian Marriage Law Postal Survey 2017 - Participation', \emph{Australian Marriage Law Postal Survey, 2017}, accessed 21 November 2022.

\url{https://www.abs.gov.au/ausstats/abs@.nsf/mf/1800.0}
}
\usage{
abs_partp2017
}
\description{
Participation results from the Australian Marriage Law Postal Survey conducted
under the Census and Statistics Act 1905. The survey asked one question:
\emph{should the law be changed to allow same-sex couples to marry?} Data presents
results at the territory/state level.
}
\details{
\code{abs_partp2017} displays the 2017 participation of an Australian age bracket
in a State/Territory and 5 variables are collected per unit:
"State/Territory", "age group", "total participants", "eligible participants",
and "participation rate". The current arrangement of the data is unsuitable
for analysis.

Eligible participants are those enrolled on the Commonwealth Electoral
Roll, or those who made a legitimate application for enrollment, and were
eligible to vote in a federal election as of 24 August 2017.
}
\keyword{datasets}
