% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_resp2017.R
\docType{data}
\name{abs_resp2017}
\alias{abs_resp2017}
\title{Australian Marriage Law Postal Survey Response Data (2017)}
\format{
A tibble with 8 rows and 6 columns
\describe{
\item{Territory/State}{State and territory totals derived from the total number of responses/eligible electors within each Federal Electoral Division assigned to that state or territory}
\item{Yes}{Count of \code{Yes} responses}
\item{Yes (\%)}{\code{Yes} responses as a proportion of total responses}
\item{No}{Count of \code{No} responses}
\item{No (\%)}{\code{No} responses as a proportion of total responses}
\item{Total}{Count of total responses}
}
}
\source{
Australian Bureau of Statistics (2017) 'Australian Marriage Law Postal Survey 2017 - Response', \emph{Australian Marriage Law Postal Survey, 2017}, accessed 21 November 2022.

\url{https://www.abs.gov.au/ausstats/abs@.nsf/Lookup/by\%20Subject/1800.0~2017~Main\%20Features~About\%20this\%20Release~9999}
}
\usage{
abs_resp2017
}
\description{
Response results from the Australian Marriage Law Postal Survey conducted
under the Census and Statistics Act 1905. The survey asked one question:
\emph{should the law be changed to allow same-sex couples to marry?} Data presents
results at the territory/state level.
}
\details{
Eligible participants are those enrolled on the Commonwealth Electoral
Roll, or those who made a legitimate application for enrollment, and were
eligible to vote in a federal election as of 24 August 2017.
}
\keyword{datasets}
