% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affordable.R
\docType{data}
\name{affordable}
\alias{affordable}
\title{Connecticut Affordable Housing by Town (2011-2021)}
\format{
A tibble with 1,686 rows and 10 columns
\describe{
\item{Town Code}{Town code}
\item{Town}{Town name}
\item{Year}{Year}
\item{2010 Census Units}{2010 Census Units}
\item{Government Assisted}{Assisted housing units or housing receiving financial assistance under any governmental program}
\item{Tenant Rental Assistance}{Rental housing occupied by persons receiving rental assistance}
\item{Single Family CHFA/ USDA Mortgages}{Ownership housing or housing currently financed by the Connecticut Housing Finance Authority and/or the U.S. Department of Agriculture}
\item{Deed Restricted Units}{Deed-restricted properties or properties with deeds containing covenants or restrictions that require such units to be sold/rented at or below prices that preserve the unit as affordable housing}
\item{Total Assisted Units}{Total assisted units}
\item{Percent Affordable}{Percent affordable units}
}
}
\source{
\if{html}{\out{<div class="sourceCode">}}\preformatted{U.S. Department of Housing
}\if{html}{\out{</div>}}

\url{https://data.ct.gov/Housing-and-Development/Affordable-Housing-by-Town-2011-2021/3udy-56vi}

\url{http://www.ct.gov/doh/cwp/view.asp?a=4513&q=530486}
}
\usage{
affordable
}
\description{
Data from the Affordable Housing Appeals Procedure List in Connecticut
provided by the U.S. Department of Housing (DOH).
}
\details{
From the authors: "DOH counts assisted housing units or housing receiving
financial assistance under any governmental program, rental housing
occupied by persons receiving rental assistance, ownership housing or
housing currently financed by the Connecticut Housing Finance Authority
and/or the U.S. Department of Agriculture, and deed-restricted properties
or properties with deeds containing covenants or restrictions that
require such dwelling unit(s) be sold or rented at or below prices that
will preserve the unit(s) as affordable housing."
}
\keyword{datasets}
