% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alana.R
\docType{data}
\name{alana}
\alias{alana}
\title{Alana's Animal Observations}
\format{
A tibble with 80 rows and 3 columns
\describe{
\item{heron}{whether a heron was observed (\code{1}) or not (\code{0})}
\item{fish}{whether a fish was observed (\code{1}) or not (\code{0})}
\item{squirrel}{whether a squirrel was observed (\code{1}) or not (\code{0})}
}
}
\usage{
alana
}
\description{
Artificial data on animal sightings over an 80-day trip.
}
\keyword{datasets}
