% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\docType{data}
\name{applications}
\alias{applications}
\title{Applicant Evaluation Scores}
\format{
A tibble with 22 rows and 5 columns
\describe{
\item{Last}{Last name of applicant}
\item{Mary}{Score given by Judge Mary}
\item{Nancy}{Score given by Judge Nancy}
\item{Olivia}{Score given by Judge Olivia}
\item{Paula}{Score given by Judge Paula}
}
}
\usage{
applications
}
\description{
Synthetic data on evaluation of applicants to a scholarship program by four judges. Each judge evaluated each applicant on a scale of 5 points.
}
\keyword{datasets}
