% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band_of_the_hour.R
\docType{data}
\name{band_of_the_hour}
\alias{band_of_the_hour}
\title{Sample character vector for practice}
\format{
A character vector
}
\source{
\url{https://welcome.miami.edu/about-um/traditions/index.html}
}
\usage{
band_of_the_hour
}
\description{
\code{band_of_the_hour} is extracted from descriptive text about
"Band of the Hour" at the University of Miami.
}
\keyword{datasets}
