% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_df.R
\docType{data}
\name{corr_tibble}
\alias{corr_tibble}
\title{Datasets for Correlation Practice}
\format{
A tibble with 2,166 rows and 3 columns
\describe{
\item{x}{x coordinate}
\item{y}{y coordinate}
\item{dataset}{corresponding dataset. Six available: "cone", "curvy", "linear", "null", "perf", "quadratic"}
}
}
\usage{
corr_tibble
}
\description{
Artificial \code{x} and \code{y} data for practicing correlation with. Six datasets are present in this tibble.
}
\keyword{datasets}
