% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csc1234.R
\docType{data}
\name{csc1234}
\alias{csc1234}
\title{Programming Course Final Grades at the College of Groundhog}
\format{
A tibble with 388 rows and 2 columns
\describe{
\item{Section}{Course section (A, B, or C)}
\item{Score}{Final course score}
}
}
\usage{
csc1234
}
\description{
Synthetic data on final grades in a programming course
at the hypothetical College of Groundhog. Final grades are
divided among three different sections (A, B, and C) of
the course taught in a typical undergraduate semester.
}
\keyword{datasets}
