% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csc_course_hw.R
\docType{data}
\name{csc_course_hw}
\alias{csc_course_hw}
\title{Computer Science Course Assessment Results}
\format{
A tibble with 116 rows and 12 columns
\describe{
\item{number}{Homework number}
\item{hw01 .. hw11}{Score for a corresponding homework (100 point scale)}
}
}
\usage{
csc_course_hw
}
\description{
Assessment results from a hypothetical introductory course
in Computer Science taught in a typical semester setting.
Results are divided across three different
tables: homeworks, labs, and quizzes.
}
\keyword{datasets}
