% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csc_course_lab.R
\docType{data}
\name{csc_course_lab}
\alias{csc_course_lab}
\title{Computer Science Course Assessment Results}
\format{
A tibble with 116 rows and 13 columns
\describe{
\item{number}{Lab number}
\item{lab01 .. lab12}{Score for a corresponding lab (100 point scale)}
}
}
\usage{
csc_course_lab
}
\description{
Assessment results from a hypothetical introductory course
in Computer Science taught in a typical semester setting.
Results are divided across three different
tables: homeworks, labs, and quizzes.
}
\keyword{datasets}
