% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csc_course_qz.R
\docType{data}
\name{csc_course_qz}
\alias{csc_course_qz}
\title{Computer Science Course Assessment Results}
\format{
A tibble with 116 rows and 4 columns
\describe{
\item{number}{Quiz number}
\item{qz01 ... qz03}{Score for a corresponding quiz (100 point scale)}
}
}
\usage{
csc_course_qz
}
\description{
Assessment results from a hypothetical introductory course
in Computer Science taught in a typical semester setting.
Results are divided across three different
tables: homeworks, labs, and quizzes.
}
\keyword{datasets}
