% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csc_labs.R
\docType{data}
\name{csc_labs}
\alias{csc_labs}
\title{Midterm and Final Scores at the University of Pluto}
\format{
A tibble with 40 rows and 3 columns
\describe{
\item{midterm}{Score on midterm exam (100 point scale)}
\item{final}{Score on the final exam (100 point scale)}
\item{section}{Lab section student was enrolled in ("F", "D", or "H")}
}
}
\usage{
csc_labs
}
\description{
Artificial data on assessment scores from a Computer Science course at the University of Pluto that had 40 enrolled students. The course was divided into three different lab sections. Data contains midterm and final results broken down by each section.
}
\keyword{datasets}
