% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\docType{data}
\name{data_types}
\alias{data_types}
\alias{bank_loans}
\alias{bands}
\title{Sample vectors for practicing data types}
\format{
A double vector

A charcter vector
}
\usage{
bank_loans

bands
}
\description{
Sample vectors for practicing operations on different data types
(e.g., element-wise operations, stringr operations, etc.)
}
\keyword{datasets}
