% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_primary.R
\docType{data}
\name{house_primary}
\alias{house_primary}
\title{U.S. House Primary Election Results (2012-2018)}
\format{
A tibble with 5,716 rows and 25 columns
}
\source{
Miller, Michael and Camberg, Nicki. Harvard Dataverse. U.S. House Primary Election Results (2012-2018) (2020).

\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/CXVMSY}

\url{https://dataverse.harvard.edu/file.xhtml?fileId=4203036&version=1.1}
}
\usage{
house_primary
}
\description{
Data on primary election results for Democratic and Republican
U.S. House candidates running in elections from 2012 to 2018.
}
\details{
For information on individual variables in this dataset, please
refer to the Codebook in the Harvard Dataverse repository
}
\keyword{datasets}
