% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longjump.R
\docType{data}
\name{longjump}
\alias{longjump}
\title{Men's Long Jump Results at the 2012 and 2016 Summer Olympics}
\format{
A tibble with 94 rows and 7 columns
\describe{
\item{year}{year of the games}
\item{event}{\code{qualifier} or \code{final}}
\item{rank}{numerical placement in the event}
\item{name}{athlete name}
\item{country}{nation the athlete represents}
\item{distance}{long jump distance (meters)}
\item{status}{qualification standard 8.10m (\code{Q}) or at least best 12 qualified (\code{q})}
}
}
\source{
Wikipedia

\url{https://en.wikipedia.org/wiki/Athletics_at_the_2012_Summer_Olympics_\%E2\%80\%93_Men\%27s_long_jump}
}
\usage{
longjump
}
\description{
Data on the results of the men's long jump qualifier and final
at the 2012 and 2016 Summer Olympic Games.
}
\keyword{datasets}
