% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mango_exams.R
\docType{data}
\name{mango_exams}
\alias{mango_exams}
\title{Assessment results at University of Mango}
\format{
A tibble with 440 rows and 4 columns
\describe{
\item{Year}{A "pseudo-year" giving the offering associated with the assessment result. It can be assumed that order is preserved in the years given.}
\item{Midterm1}{Score on the first midterm exam (100 point scale)}
\item{Midterm2}{Score on the second midterm exam (100 point scale)}
\item{Final}{Score on the final exam (100 point scale)}
}
}
\usage{
mango_exams
}
\description{
Artificial data on assessment scores from four different offerings of a course taught at the University of Mango. The course contained two midterm assessments and a final exam administered at the end of the semester.
}
\keyword{datasets}
