% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moby_dick_counts.R
\docType{data}
\name{moby_dick_counts}
\alias{moby_dick_counts}
\title{Sample tibble containing word counts for practice}
\format{
A tibble with 150 rows and 5 columns
\describe{
\item{chapters}{Main text of the chapter}
\item{chapter_num}{Chapter number}
\item{whale}{Word count for the word "whale"}
\item{sea}{Word count for the word "sea"}
\item{ahab}{Word count for the word "ahab"}
}
}
\source{
Project Gutenberg

\url{https://www.gutenberg.org}
}
\usage{
moby_dick_counts
}
\description{
A tibble containing word counts per chapter for three different
words. Extracted from Moby Dick on Project Gutenberg.
}
\keyword{datasets}
