% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mystery_flights.R
\docType{data}
\name{mystery_flights}
\alias{mystery_flights}
\title{Rearranged Flights Data}
\format{
A tibble with 336,776 rows and 20 columns
}
\usage{
mystery_flights
}
\description{
Rearranged data for all flights that departed NYC (i.e. JFK, LGA or EWR) in 2013. Sourced from \code{flights} in the \code{nycflights13} package. This dataset is meant to showcase issues with convenience samples (all flights with an even row index are early departures).
}
\details{
Please see the \code{nycflights13} package for information on this dataset.
}
\keyword{datasets}
