% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluto.R
\docType{data}
\name{pluto}
\alias{pluto}
\title{Major Preferences at College of Pluto}
\format{
A tibble with 12 rows and 4 columns
\describe{
\item{Major}{Major}
\item{Gender}{Male or Female}
\item{Applied}{Count of applicants applied to major}
\item{Accepted}{Count of applicants accepted to major}
}
}
\usage{
pluto
}
\description{
Data on preferences for six different majors at the hypothetical College of Pluto
}
\keyword{datasets}
