% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_fl.R
\docType{data}
\name{scorecard_fl}
\alias{scorecard_fl}
\title{Florida College Scorecard (2022)}
\format{
A tibble with 395 rows and 2,989 columns
}
\source{
U.S. Department of Education

\url{https://collegescorecard.ed.gov/data/}

\url{https://collegescorecard.ed.gov/data/glossary/}

\url{https://collegescorecard.ed.gov/data/documentation/}
}
\usage{
scorecard_fl
}
\description{
Institution-level data sourced from the U.S. Department of Education
College Scorecard. Contains aggregate data for Florida-based institutions.
Includes information on institutional characteristics, enrollment, student aid,
costs, and student outcomes for 1996-97 through 2020-21.
}
\details{
Please refer to the data dictionary in the Scorecard Glossary for information on
specific variables in this dataset.
}
\keyword{datasets}
