% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_salary.R
\docType{data}
\name{sf_salary}
\alias{sf_salary}
\title{San Francisco City Employee Salary Data}
\format{
A tibble with 148,654 rows and 13 columns
\describe{
\item{Id}{Row number}
\item{EmployeeName}{Employee name}
\item{JobTitle}{Job title of employee}
\item{BasePay}{Base pay of employee}
\item{OvertimePay}{Overtime pay of employee}
\item{OtherPay}{Other pay of employee}
\item{Benefits}{Employee benefits}
\item{TotalPay}{Total pay compensation}
\item{TotalPayBenefits}{Employee total pay benefits}
\item{Year}{Year}
\item{Notes}{Notes corresponding to employee}
\item{Agency}{Reporting agency}
\item{Status}{Status information}
}
}
\source{
Kaggle and Transparent California

\url{https://transparentcalifornia.com/salaries/san-francisco/}

\url{https://www.kaggle.com/datasets/kaggle/sf-salaries}
}
\usage{
sf_salary
}
\description{
Data on the names, job titles, and compensation for San Francisco
city employees on an annual basis from 2011 to 2014.
}
\keyword{datasets}
