% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemp_usda.R
\docType{data}
\name{unemp_usda}
\alias{unemp_usda}
\title{Unemployment and median household income for the U.S., States, and counties, 2000-2021}
\format{
A tibble with 3,277 rows and 96 columns
\describe{
\item{FIPS_code}{State-county Federal Information Processing Standards (FIPS) code}
\item{State}{State abbreviation}
\item{Area_name}{State or county name}
\item{Rural_urban_continuum_code_2013}{Rural-urban Continuum Code in 2013}
\item{Urban_influence_code_2013}{Urban Influence Code in 2013}
\item{Metro_2013}{\code{0}=Nonmetro \code{1}=Metro (2013 OMB Metropolitan Area delineation)}
\item{Civilian_labor_force_\code{X}}{Civilian labor force annual average in year \code{X}}
\item{Employed_\code{X}}{Number employed annual average in year \code{X}}
\item{Unemployed_\code{X}}{Number unemployed annual average in year \code{X}}
\item{Unemployment_rate_\code{X}}{Unemployment rate in year \code{X}}
\item{Median_Household_Income_2019}{Estimate of median household income in 2020}
\item{Med_HH_Income_Percent_of_State_Total_2020}{County household median income as a percent of the State total median household income in 2020}
}
}
\source{
U.S. Department of Agriculture

\url{https://www.ers.usda.gov/data-products/county-level-data-sets/}
}
\usage{
unemp_usda
}
\description{
County-level data on annual average unemployment and median
household income in the U.S. Derived from the U.S. Department
of Agriculture ERS (Economic Research Service).
}
\keyword{datasets}
